+++ releaseherald/releaseherald/configuration.py
@@ -11,21 +11,6 @@ DEFAULT_FRAGMENTS_DIR = Path("news_fragments")
 
 DEFAULT_VERSION_TAG_PATTERN = re.compile(r"(?P<version>(\d*)\.(\d*)\.(\d*))")
 
-
-class SubmoduleConfig(BaseModel):
-    name: str
-    display_name: Optional[str] = None
-    news_fragments_directory: Path = DEFAULT_FRAGMENTS_DIR
-    version_tag_pattern: Pattern = DEFAULT_VERSION_TAG_PATTERN
-
-    @root_validator
-    def default_display_name(cls, values):
-        values = values.copy()
-        display_name = values.get("display_name")
-        values["display_name"] = display_name or values["name"]
-        return values
-
-
 DefaultOptionsCallable = Callable[[Dict[str, Any]], None]
 
 
@@ -41,7 +26,6 @@ class Configuration(BaseModel):
     last_tag: str = ""
     latest: bool = False
     update: bool = True
-    submodules: List[SubmoduleConfig] = []
 
     class Config:
         paths_to_resolve: List[str] = ["template", "news_file"]
